+++ src/main/kotlin/dev/d1s/linda/entity/Redirect.kt
@@ -16,25 +16,19 @@
 
 package dev.d1s.linda.entity
 
+import dev.d1s.linda.entity.common.Identifiable
 import dev.d1s.linda.entity.utmParameter.UtmParameter
-import org.hibernate.annotations.GenericGenerator
-import java.time.Instant
+import org.hibernate.Hibernate
 import javax.persistence.*
 
 @Entity
 @Table(name = "redirect")
 data class Redirect(
+
     @ManyToOne(cascade = [CascadeType.MERGE])
     var shortLink: ShortLink
-) : Identifiable {
-    @Id
-    @Column
-    @GeneratedValue(generator = "system-uuid")
-    @GenericGenerator(name = "system-uuid", strategy = "uuid")
-    override var id: String? = null
 
-    @Column
-    override var creationTime: Instant? = null
+) : Identifiable() {
 
     @ManyToMany
     @JoinTable(
@@ -44,34 +38,17 @@ data class Redirect(
     )
     var utmParameters: MutableSet<UtmParameter> = mutableSetOf()
 
-    @PrePersist
-    fun setCreationTime() {
-        creationTime = Instant.now()
-    }
-
     override fun equals(other: Any?): Boolean {
         if (this === other) return true
-        if (other !is Redirect) return false
+        if (other == null || Hibernate.getClass(this) != Hibernate.getClass(other)) return false
+        other as Redirect
 
-        if (shortLink != other.shortLink) return false
-        if (id != other.id) return false
-        if (creationTime != other.creationTime) return false
-
-        return true
+        return id != null && id == other.id
     }
 
-    override fun hashCode(): Int {
-        var result = shortLink.hashCode()
-        result = 31 * result + (id?.hashCode() ?: 0)
-        result = 31 * result + (creationTime?.hashCode() ?: 0)
-        return result
-    }
+    override fun hashCode(): Int = javaClass.hashCode()
 
     override fun toString(): String {
-        return "Redirect(" +
-                "shortLink=$shortLink, " +
-                "id=$id, " +
-                "creationTime=$creationTime, " +
-                "utmParameters=$utmParameters)"
+        return "Redirect(shortLink=$shortLink, utmParameters=$utmParameters)"
     }
 }
